requirements.txt
aniso8601==8.0.0
click==7.1.2
enum-compat==0.0.3
Flask==1.1.2
Flask-RESTful==0.3.8
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
yeelight==0.5.3
